+++ app/src/main/java/app/books/tanga/data/category/CategoryRepository.kt
@@ -1,6 +1,7 @@
-package app.books.tanga.data
+package app.books.tanga.data.category
 
-import app.books.tanga.domain.categories.Category
+import app.books.tanga.data.FirestoreDatabase
+import app.books.tanga.entity.Category
 import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
