+++ app/src/main/java/app/books/tanga/entity/Error.kt
@@ -1,3 +1,3 @@
-package app.books.tanga.domain
+package app.books.tanga.entity
 
 class OperationError(message: String): Throwable(message)
