+++ app/src/main/java/app/books/tanga/feature/home/HomeViewModel.kt
@@ -3,9 +3,9 @@ package app.books.tanga.feature.home
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.books.tanga.data.UserRepository
-import app.books.tanga.domain.favorites.FavoriteInteractor
-import app.books.tanga.domain.summary.SummaryInteractor
+import app.books.tanga.data.user.UserRepository
+import app.books.tanga.feature.library.FavoriteInteractor
+import app.books.tanga.feature.summary.SummaryInteractor
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
