+++ app/src/main/java/app/books/tanga/feature/library/LibraryViewModel.kt
@@ -3,7 +3,6 @@ package app.books.tanga.feature.library
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.books.tanga.common.ui.ProgressState
-import app.books.tanga.domain.favorites.FavoriteInteractor
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
