+++ app/src/main/java/app/books/tanga/feature/search/SearchUiContract.kt
@@ -3,8 +3,8 @@ package app.books.tanga.feature.search
 import androidx.annotation.DrawableRes
 import app.books.tanga.common.ui.ProgressState
 import app.books.tanga.core_ui.icons.TangaIcons
-import app.books.tanga.domain.categories.Category
-import app.books.tanga.domain.categories.PredefinedCategory
+import app.books.tanga.entity.Category
+import app.books.tanga.entity.PredefinedCategory
 import app.books.tanga.feature.summary.SummaryUi
 
 data class SearchUiState(
