+++ app/src/main/java/app/books/tanga/feature/splash/SplashViewModel.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.books.tanga.navigation.NavigationScreen
 import app.books.tanga.data.preferences.DefaultPrefDataStoreRepository
-import app.books.tanga.domain.session.SessionManager
-import app.books.tanga.domain.session.SessionState
+import app.books.tanga.session.SessionManager
+import app.books.tanga.session.SessionState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
