+++ app/src/main/java/app/books/tanga/session/SessionState.kt
@@ -1,4 +1,4 @@
-package app.books.tanga.domain.session
+package app.books.tanga.session
 
 sealed class SessionState {
     data class LoggedIn(val sessionId: SessionId): SessionState()
